requirements.txt
customtkinter
ffmpeg-python
Pillow
tkinterdnd2
